chore: commit pre-existing audit docs, backend fixes, tests, and config
Includes backend type-safety fixes (converter, formatos, sellador, technical
reports), Electron IPC handler updates, frontend test-setup and RenameCard
mapping improvements, shared html-sanitizer module, CI workflow, audit
documentation, scratch diagnostic artifacts, and .gitignore updates for
build artifacts.

+++ scripts/generate_brand_assets.py
@@ -18,7 +18,6 @@ from __future__ import annotations
 
 import base64
 import io
-import shutil
 import struct
 from pathlib import Path
 
@@ -71,7 +70,7 @@ def load_and_trim(path: Path) -> Image.Image:
 def resize_to_height(img: Image.Image, height: int) -> Image.Image:
     w, h = img.size
     scale = height / h
-    new_w = max(1, int(round(w * scale)))
+    new_w = max(1, round(w * scale))
     return img.resize((new_w, height), Image.Resampling.LANCZOS)
